add confirm alert when deleting list

+++ frontend/src/components/ListHome.js
@@ -24,6 +24,9 @@ const ListHome = (props) => {
     };
 
     function deleteList(url) {
+        if (window.confirm("Delete this item?") === false)
+            return;
+
         axios.delete(url, { headers: { 'Authorization': `Token ${JSON.parse(localStorage.getItem('tokens'))['token']}` }})
             .then(res => getLists())
             .catch(err => console.log(err));
